currency conversion and controls

+++ CoinConversion/CoinConversion/Services/Conversion/Entity/CurrenciesConversion.swift
@@ -13,7 +13,7 @@ struct CurrenciesConversion: Codable {
     let success: Bool
     let terms: String
     let privacy: String
-    let timestamp: Int
+    let timestamp: Double
     let source: String
     let quotes: [String: Double]
